All users CRUD OK

+++ app/Http/Controllers/UsersController.php
@@ -7,6 +7,8 @@ use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Facades\Hash;
 use App\User;
 use App\Org;
+use App\UserTransactions;
+use Auth;
 use Illuminate\Http\Request;
 
 class UsersController extends Controller
@@ -40,15 +42,26 @@ class UsersController extends Controller
      */
     public function store(StoreUser $request)
     {
-      $validatedData = $request->validated();
+      $validatedData = $request->except(['credit','paymentMethod']);
       if( !$request->password ){$validatedData['password']=env('DEFAULT_PASSWORD','ledamcha');}
       if( $validatedData['avatar'] == '' ){$validatedData['avatar'] = '/placeholders/avatar-male.png';}
       $validatedData['password'] = Hash::make($validatedData['password']);
       $user = User::create($validatedData);
+
+      $userTrans = new UserTransactions;
+      $userTrans->user_id = $user->id;
+      $userTrans->credit = $request->credit;
+      $userTrans->paymentMethod = $request->paymentMethod;
+      $userTrans->save();
+
       Session::flash('message', env("SAVE_SUCCESS_MSG","User saved succesfully!"));
 
      if( $request->type == 3 ){
         return redirect(route('admin.index'));
+      }elseif ( $request->type == 4 ) {
+        return redirect(route('suppliers.index'));
+      }elseif ( $user->type == 2 ) {
+        return redirect(route('customers.index'));
       }
       return redirect(route('staff.index'));
     }
@@ -89,6 +102,14 @@ class UsersController extends Controller
       if( !$request->password ){$validatedData['password']=env('DEFAULT_PASSWORD','ledamcha');}
       $validatedData['password'] = Hash::make($validatedData['password']);
       $user->update($validatedData);
+      if( $request->credit > 0 ){
+        $userTrans = new UserTransactions;
+        $userTrans->user_id = $user->id;
+        $userTrans->credit = $request->credit;
+        $userTrans->paymentMethod = $request->paymentMethod;
+        $userTrans->save();
+      }
+
       Session::flash('message', env("SAVE_SUCCESS_MSG","User updated succesfully!"));
       return back();
     }
@@ -101,11 +122,20 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
+      if( $user->id == Auth::ID() ){
+        Session::flash('error', env("SAVE_SUCCESS_MSG","You cannot delete yourself!"));
+        return back();
+      }
+      UserTransactions::where('user_id',$user->id)->delete();
       $user->delete();
       Session::flash('message', env("SAVE_SUCCESS_MSG","User deleted succesfully!"));
 
       if( $user->type == 3 ){
          return redirect(route('admin.index'));
+       }elseif ( $user->type == 4 ) {
+         return redirect(route('suppliers.index'));
+       }elseif ( $user->type == 2 ) {
+         return redirect(route('customers.index'));
        }
 
        return redirect(route('staff.index'));
